% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kemma_vorst_arithmetic.R
\name{print.kemma_vorst_arithmetic}
\alias{print.kemma_vorst_arithmetic}
\title{Print Method for Kemma-Vorst Arithmetic Results}
\usage{
\method{print}{kemma_vorst_arithmetic}(x, ...)
}
\arguments{
\item{x}{Object of class "kemma_vorst_arithmetic"}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Kemma-Vorst Arithmetic Results
}
