% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kemma_vorst_arithmetic.R
\name{summary.kemma_vorst_arithmetic}
\alias{summary.kemma_vorst_arithmetic}
\title{Summary Method for Kemma-Vorst Arithmetic Results}
\usage{
\method{summary}{kemma_vorst_arithmetic}(object, ...)
}
\arguments{
\item{object}{Object of class "kemma_vorst_arithmetic"}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary Method for Kemma-Vorst Arithmetic Results
}
